complete 'happy' path simulator

+++ ios/client/Sm02Client/Sources/Sm02Client/RemoteServer.swift
@@ -3,9 +3,9 @@ public struct RemoteServer: Codable {
 
   public let ssid: String
   public let ip: String
-  public let code: [Int]
+  public let code: [UInt8]
 
-  public init (ssid: String, ip: String, code: [Int]) {
+  public init (ssid: String, ip: String, code: [UInt8]) {
     self.ssid = ssid
     self.ip = ip
     self.code = code
